requirements.txt
flask
pandas
numpy
google-generativeai